gitserver: Fix infinite clone loop for repo duplicates
This is admittedly a bit of a hack, but this cost us thousands of dollars on cloud before so I don't feel too terrible about this.
The real solution will require to track repos by an actually unique identifier that is immutable, but that's much more work.

After the various fixes made to repo name normalization, there's less instances of "accidentally updating the wrong DB entry".
In fact, after the previous PR everything already works alright.
Except on a restart, where gitserver scans all repos and updates their clone status.

To work around this, we now exclude repos from state updates in the initial run that have another undeleted repo with the same name.

This is not a complete solution, but addresses all the happy path cases I could find where a repo was incorrectly recloned.

## Test plan

Recreated the scenario where repo A is deleted, and repo B has taken the name of repo A in the database.
On main, we regularly delete this repo and reclone it. On this branch, this doesn't happen.
Also adjusted tests for the store method.